Move error class to base.py
